networking tests: fix ssh auth, clean up code
- Fix SSH auth using password. Paramiko can get confused when using
  password authentication while having SSH keys in the SSH agent. In
  case the test suite is being run using password based authentication,
  the options `allow_agent=False` and `look_for_key=False` prevent
  paramiko from trying to get keys from the SSH agent, allowing password
  based authentication to proceed without error.

- Clean up duplicated fixture `vlan_id` and `vlan_nic`. These pytest
  fixtures were copy-and-pasted into several places, so to simplify the
  code they are consolidated in a single place where other network
  fixtures can be kept in the future as well.
